Extract tag-priority logic from PromptGenerator into lib/promptGeneration.ts
Moves the pure helpers (shuffleArray, TAG_GROUPS, classifyTag,
getTagPriority, makeExcludeFilter) plus a new sortByPriority helper into
a testable module with no React dependency. handleGenerate — previously
stateful, reading ~10 pieces of component state directly and writing via
setGeneratedTags — is refactored into a pure generatePrompt(modelTags,
options) function; the component now just calls it and applies the
result. The two useEffects that live-update subject type and breast size
without re-rolling now also import classifyTag/sortByPriority from the
same module, instead of duplicating that logic locally.

This is a refactor, not a behavior change. Verified via a real browser
smoke test against a running container: generated a prompt, confirmed
priority ordering (subject → solo → nudity → character → breasts →
hair → clothing → misc) matches the original algorithm, and confirmed
switching subject type (1girl → 1boy) still live-updates in place
without re-rolling the rest of the prompt.
